+++ controlplane/src/db/schema.ts
@@ -14,6 +14,7 @@ import {
   customType,
   real,
 } from 'drizzle-orm/pg-core';
+import { FeatureIds } from '../types/index.js';
 
 // JSON/JSONB custom types to workaround insert bug
 // Should not be used with other drivers than postgres-js
@@ -236,7 +237,9 @@ export const targets = pgTable(
     labels: text('labels').array(),
     organizationId: uuid('organization_id')
       .notNull()
-      .references(() => organizations.id),
+      .references(() => organizations.id, {
+        onDelete: 'cascade',
+      }),
     createdBy: uuid('created_by').references(() => users.id, {
       onDelete: 'set null',
     }),
@@ -529,7 +532,9 @@ export const apiKeys = pgTable(
       .references(() => users.id),
     organizationId: uuid('organization_id')
       .notNull()
-      .references(() => organizations.id),
+      .references(() => organizations.id, {
+        onDelete: 'cascade',
+      }),
     name: text('name').notNull(),
     key: text('key').unique().notNull(),
     lastUsedAt: timestamp('last_used_at', { withTimezone: true }),
@@ -556,7 +561,7 @@ export const graphApiTokens = pgTable(
     id: uuid('id').notNull().primaryKey().defaultRandom(),
     organizationId: uuid('organization_id')
       .notNull()
-      .references(() => organizations.id),
+      .references(() => organizations.id, { onDelete: 'cascade' }),
     federatedGraphId: uuid('federated_graph_id')
       .notNull()
       .references(() => federatedGraphs.id, { onDelete: 'cascade' }),
@@ -576,7 +581,7 @@ export const graphRequestKeys = pgTable('graph_request_keys', {
   id: uuid('id').notNull().primaryKey().defaultRandom(),
   organizationId: uuid('organization_id')
     .notNull()
-    .references(() => organizations.id),
+    .references(() => organizations.id, { onDelete: 'cascade' }),
   federatedGraphId: uuid('federated_graph_id')
     .notNull()
     // Only one request key per federated graph
@@ -608,7 +613,6 @@ export const organizationBilling = pgTable(
         onDelete: 'cascade',
       }),
     plan: text('plan'),
-    email: text('email'),
     stripeCustomerId: text('stripe_customer_id'),
   },
   (t) => {
@@ -620,7 +624,7 @@ export const organizationBilling = pgTable(
 );
 
 export type Feature = {
-  id: string;
+  id: FeatureIds;
   description?: string;
   limit?: number;
 };