+++ studio/CHANGELOG.md
@@ -4,6 +4,22 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.47.3](https://github.com/wundergraph/cosmo/compare/studio@0.47.2...studio@0.47.3) (2024-01-01)
+
+### Bug Fixes
+
+* reset plan after subscription del, bill for upgrade immediately ([b649383](https://github.com/wundergraph/cosmo/commit/b649383446bf8a6187a61d795d4237604f401c87)) (@StarpTech)
+
+## [0.47.2](https://github.com/wundergraph/cosmo/compare/studio@0.47.1...studio@0.47.2) (2024-01-01)
+
+### Bug Fixes
+
+* **payment:** handle customer deletion, cancel subs when org is deleted ([#401](https://github.com/wundergraph/cosmo/issues/401)) ([b58809d](https://github.com/wundergraph/cosmo/commit/b58809dca894a6aab879d1750bfd06a608da207f)) (@StarpTech)
+
+## [0.47.1](https://github.com/wundergraph/cosmo/compare/studio@0.47.0...studio@0.47.1) (2023-12-31)
+
+**Note:** Version bump only for package studio
+
 # [0.47.0](https://github.com/wundergraph/cosmo/compare/studio@0.46.0...studio@0.47.0) (2023-12-29)
 
 ### Features